requirements.txt
python-telegram-bot
requests
python-dotenv
PySocks
selenium
twilio